@@ -32,7 +32,7 @@
     "@swc/core": "^1.4.2",
     "@types/node": "^20.11.24",
     "tsup": "^8.0.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.7.2"
   },
   "publishConfig": {
     "access": "public"
