@@ -1,9 +1,11 @@
-const withNextra = require("nextra")({
+import nextra from "nextra";
+
+const withNextra = nextra({
   theme: "nextra-theme-docs",
   themeConfig: "./theme.config.tsx",
 });
 
-module.exports = withNextra({
+export default withNextra({
   output: "export",
   images: {
     unoptimized: true,
