@@ -1,5 +1,7 @@
-{
+const meta = {
   "deploy-vercel": "Deploy to Vercel",
   "deploy-netlify": "Deploy to Netlify",
-  "deploy-render": "Deploy to Render"
-}
+  "deploy-render": "Deploy to Render",
+};
+
+export default meta;
